update shell script

+++ env/runos.sh
@@ -1,5 +1,7 @@
 #!/bin/zsh
 
+debug=$1
+
 if [ -e kernel.s ]; then
   nasm boot.s -o boot.bin -l boot.lst
   nasm kernel.s -o kernel.bin -l kernel.lst
@@ -8,8 +10,11 @@ else
   nasm boot.s -o boot.img -l boot.lst
 fi
 
-rm boot.bin kernel.bin kernel.lst boot.lst
+if [ ! -z "$debug" ] && [ "$debug" = "debug" ]; then 
+  echo "debug"
+else 
+  echo "not debug"
+  rm boot.bin kernel.bin kernel.lst boot.lst
+fi
 
-bochs -q -f ../../env/.bochsrc -rc ../../env/cmd.init
-
-rm boot.img 
+bochs -q -f ../../env/.bochsrc -rc ../../env/cmd.init;
